Rename `strongs.rs` and remove unmaintained tests
