+++ circ/src/strong.rs
